+++ source/main/java/org/opensagetv/vibe/tmdb/TmdbMetadataService.java
@@ -37,4 +37,6 @@ public interface TmdbMetadataService extends Closeable {
 
   Optional<TmdbArtworkConfiguration> getCachedArtworkConfiguration()
       throws IOException, SQLException;
+
+  LibraryEnrichmentService getLibraryEnrichmentService();
 }